chore: Split trie and builder to their own files.
