+++ src/trie/mod.rs
@@ -1,3 +1,5 @@
 mod trie;
+mod trie_builder;
 
-pub use trie::{Trie, TrieBuilder};
+pub use trie::Trie;
+pub use trie_builder::TrieBuilder;
